gallery/backend/pyproject.toml
[project]
name = "gallery-backend"
version = "0.1.0"
description = "Gallery backend API for storing and managing images with metadata"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0,<7.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.0.0",
    "boto3>=1.35.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "gunicorn>=21.0.0",
    "drf-spectacular>=0.27.0",
]
